# packages/api-service/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Turnstile API",
    "description": "Token bridge system for Ethereum L1 and Aztec L2 networks",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {},
...
  },
  "paths": {
    "/tokens": {
      "get": {
        "summary": "List all tokens",
        "tags": [
          "Tokens"
        ],
        "description": "Get a paginated list of all tokens",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Maximum number of items to return (1-1000)"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false,
            "description": "Cursor for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number"
                          },
                          "symbol": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "decimals": {
                            "type": "number"
                          },
                          "address": {
                            "type": "string",
                            "pattern": "^0x[a-fA-F0-9]{64}$"
                          }
                        },
                        "required": [
                          "symbol",
                          "name",
                          "decimals",
                          "address"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "limit": {
                          "type": "number"
                        },
                        "cursor": {
                          "type": "number"
                        },
                        "nextCursor": {
                          "type": "number"
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "limit",
                        "hasMore"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "pagination"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/tokens/{address}": {
      "get": {
        "summary": "Get token by address",
        "tags": [
          "Tokens"
        ],
        "description": "Get a token by its address",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "address",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "symbol": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "decimals": {
                      "type": "number"
                    },
                    "address": {
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    }
                  },
                  "required": [
                    "symbol",
                    "name",
                    "decimals",
                    "address"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Development server"
    }
  ],
...
}